Created : component `MultiItemInput`
